Pricing and a slightly revised homepage.
Snuck the word "server" onto the homepage (for the n00bs and the' g00gs) but still need to revise the features copy. Also, not 100% happy with the quotes on the pricing page, they still don't feel "legit" but that may be a function of them not being a lot or amazing testimonials.