`assert_unauthenticated_response` tests for FailureApp.
Before it just tested the *result* of FailureApp rather than wether it was actually called. Because the result changed depending on the format of the request (html or xml) good tests would fail the assertion.

The fix is a bit hacky but it works and the hack is documented. If it breaks half the codebase will also fail so we'll know about it pretty quickly.